refactor: moved FormBuilderPlayground to its own folder
