+++ react-client/src/routes/playground/playgrounds/index.ts
@@ -10,7 +10,7 @@ export * from "./ConfirmDeletionDialogPlayground";
 export * from "./DialogPlayground";
 export * from "./DragAndDropZonePlayground";
 export * from "./DropdownMenuPlayground";
-export * from "./FormBuilderPlayground";
+export * from "./formBuilder/FormBuilderPlayground";
 export * from "./MyAccountPagePlayground";
 export * from "./OnboardingDialogPlayground";
 export * from "./PetCardRecordPlayground";
